// src/interfaces/interfaces.ts
export interface DataProduct {
    _id: string;
    nameProduct: string;
    marca: string;
    priceFactoryProduct: number;
    priceStoreProduct: number;
    imageProduct: string;
    imagePath: string;
}

export interface Product {
    nameProduct: string,
    marca: string,
    priceFactoryProduct: number,
    priceStoreProduct: number,
    imageProduct: string,
    imagePath: string;
}
  